Move Network to `o.b.base`
