+++ core/src/main/java/org/bitcoinj/base/Network.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.bitcoinj.utils;
+package org.bitcoinj.base;
 
 import java.util.Arrays;
