+++ core/src/main/java/org/bitcoinj/params/AbstractBitcoinNetParams.java
@@ -32,7 +32,7 @@ import org.bitcoinj.protocols.payments.PaymentProtocol;
 import org.bitcoinj.store.BlockStore;
 import org.bitcoinj.store.BlockStoreException;
 import org.bitcoinj.base.utils.MonetaryFormat;
-import org.bitcoinj.utils.Network;
+import org.bitcoinj.base.Network;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
