+++ core/src/main/java/org/bitcoinj/wallet/DefaultRiskAnalysis.java
@@ -26,7 +26,7 @@ import org.bitcoinj.core.TransactionInput;
 import org.bitcoinj.core.TransactionOutput;
 import org.bitcoinj.crypto.TransactionSignature;
 import org.bitcoinj.script.ScriptChunk;
-import org.bitcoinj.utils.Network;
+import org.bitcoinj.base.Network;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
