+++ core/src/main/java/org/bitcoinj/wallet/KeyChainGroupStructure.java
@@ -20,7 +20,7 @@ import org.bitcoinj.base.ScriptType;
 import org.bitcoinj.core.NetworkParameters;
 import org.bitcoinj.crypto.ChildNumber;
 import org.bitcoinj.crypto.HDPath;
-import org.bitcoinj.utils.Network;
+import org.bitcoinj.base.Network;
 
 /**
  *  Defines a structure for hierarchical deterministic wallets.
