+++ core/src/test/java/org/bitcoinj/core/LegacyAddressTest.java
@@ -28,7 +28,7 @@ import org.bitcoinj.script.Script;
 import org.bitcoinj.base.ScriptType;
 import org.bitcoinj.script.ScriptBuilder;
 import org.bitcoinj.script.ScriptPattern;
-import org.bitcoinj.utils.Network;
+import org.bitcoinj.base.Network;
 import org.junit.Test;
 
 import java.io.BufferedReader;
